requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.1.0
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
bullet==2.2.0
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chex
click==8.1.3
cloudpickle==2.2.0
cmake==3.26.3
cmx==0.0.18
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
dataclasses==0.6
debugpy==1.6.3
decorator==4.4.2
dm-tree==0.1.7
docker-pycreds==0.4.0
entrypoints==0.4
etils==0.9.0
exceptiongroup==1.1.0
executing==1.2.0
fasteners==0.18
filelock==3.11.0
flatbuffers==23.5.26
flax==0.6.2
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.5
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.1
gym==0.26.2
gym-deepmindlab==0.1.2
gym-fetch==0.2.2
gym-notices==0.0.8
gymnasium==0.27.0
gymnasium-notices==0.0.1
h5py==3.8.0
hydra-core==1.3.0
hydra-joblib-launcher==1.2.0
hydra-submitit-launcher==1.2.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.8
imbalanced-learn==0.10.0
imblearn==0.0
importlib-metadata==5.1.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.17.1
ipython==8.6.0
ipywidgets==8.0.2
jax==0.3.25
jax-jumpy==1.0.0
jaxlib @ https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.3.25+cuda11.cudnn805-cp310-cp310-manylinux2014_x86_64.whl#sha256=e44cc7905aa6fcdad13b95a968eb36a0adcb01d173b40a22afd88869c1b898fa
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-widgets==3.0.3
keras==2.13.1
kiwisolver==1.4.4
labmaze==1.0.6
libclang==16.0.0
lit==16.0.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.4
mujoco==2.3.3
mujoco-py==2.1.2.14
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu11==2.14.3
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax
orbax-checkpoint
packaging==21.3
pandas==1.5.2
parso==0.8.3
patchelf==0.17.0.0
pathtools==0.1.2
PettingZoo==1.22.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
pluggy==1.0.0
proglog==0.1.10
prompt-toolkit==3.0.33
protobuf==4.24.2
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.13.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pytest==7.2.0
pytest-instafail==0.5.0
pytest-warnings==0.3.1
python-dateutil==2.8.2
pytinyrenderer==0.0.14
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==11.2.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
sentry-sdk==1.19.1
setproctitle==1.3.2
Shimmy==0.2.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.1
submitit==1.4.5
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
tensordict==0.1.0
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.18.0
tensorstore==0.1.41
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch-tb-profiler==0.4.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trimesh==3.16.4
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.14.2
wcwidth==0.2.5
Werkzeug==2.3.4
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.10.0
